+++ tools/downloader/setup.cfg
@@ -30,8 +30,8 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    omz_converter = open_model_zoo.model_tools.converter:main
-    omz_data_downloader = open_model_zoo.model_tools.data_downloader:main
-    omz_downloader = open_model_zoo.model_tools.downloader:main
-    omz_info_dumper = open_model_zoo.model_tools.info_dumper:main
-    omz_quantizer = open_model_zoo.model_tools.quantizer:main
+    omz_converter = openvino.model_zoo.converter:main
+    omz_data_downloader = openvino.model_zoo.data_downloader:main
+    omz_downloader = openvino.model_zoo.downloader:main
+    omz_info_dumper = openvino.model_zoo.info_dumper:main
+    omz_quantizer = openvino.model_zoo.quantizer:main
