+++ tools/downloader/src/openvino/model_zoo/_common.py
