+++ tools/downloader/src/openvino/model_zoo/converter.py
@@ -21,7 +21,7 @@ import sys
 
 from pathlib import Path
 
-from open_model_zoo.model_tools import (
+from openvino.model_zoo import (
     _configuration, _common, _concurrency, _reporting,
 )
